+++ .gitignore
@@ -1,1 +1,4 @@
 test_export
+dist
+package-lock.json
+node_modules
